// src/app/api/models/theme-dto.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface ThemeDto {
  accentColor?: string;
  active?: boolean;
  allowCustomCss?: boolean;
  backgroundColor?: string;
  buttonTextColor?: string;
  code?: string;
  customCss?: string;
  faviconUrl?: string;
  fontFamily?: string;
  id?: number;
  linkColor?: string;
  logoUrl?: string;
  name?: string;
  primaryColor?: string;
  propertiesJson?: string;
  secondaryColor?: string;
  tenantId?: number;
  textColor?: string;
  themeMode?: 'LIGHT' | 'DARK' | 'AUTO';
  version?: number;
  warnColor?: string;
}